refactored test cases
